+++ src/Fabulous.MauiControls/Views/Controls/Span.fs
@@ -70,7 +70,7 @@ type SpanModifiers =
     static member inline font
         (
             this: WidgetBuilder<'msg, #ISpan>,
-            ?size: double,
+            ?size: float,
             ?namedSize: NamedSize,
             ?attributes: FontAttributes,
             ?fontFamily: string
